sil: Implement basic parser, remove a lot of lifetimes
It became somewhat difficult to read/write code with explicit lifetimes
applied throughout everything, so instead we use pointers for some of
the things and pretend strings are static. These objects should not live
longer than the file they built from, so this should still be safe as long
as we keep that in mind.